Adds early support to change local user in app
This commit adds the base needed to change user info in the app. It is require for Login/Register.
Notice: this doesn't deal yet with how to register/login. It only calls a setter from main app component, so we could change it when we deal with the login logic.