fix dynamic change of user and selected user chat, and cleanup of code
Fix an issue when you login with your user and password, nothing changes on screen. This was because the change does happens behind the scenes, but the front doesn't know its need to reprint the elements with new data.
Also did some cleanup and add few comments where it seems necessary for code readability.